+++ CHANGELOG.md
@@ -5,10 +5,8 @@
 ### Bugs Fixed
 
 * Fixed an issue that could cause `Conn.connReader()` to become blocked in rare circumstances.
-
-### Bugs Fixed
-
 * Fixed an issue that could cause outgoing transfers to be rejected by some brokers due to out-of-sequence delivery IDs.
+* Fixed an issue that could cause senders and receivers within the same session to deadlock if the receiver was configured with `ReceiverSettleModeFirst`.
 
 ### Other Changes